Create README.md
This program offers a way to process and interpret the Enron dataset. It mostly look at features helping to classify persons involved in the Enron's fraud. To run this program you first need to run the "../tools/startup.py" script to download the email corpus and the needed python libraries.